requirements.txt
Flask>=0.10
flask-cors==3.0.4
PyYAML>=3.0
jsonschema==2.5.1
jsonref==0.1
